Extract `Filter` out to the helpers package
