+++ tui/helpers/utils.go
@@ -1,0 +1,15 @@
+package helpers
+
+// UTILITIES
+// ---------
+
+// Filter filters out empty strings from a slice of strings
+func Filter(s ...string) []string {
+	var filtered []string
+	for _, v := range s {
+		if v != "" {
+			filtered = append(filtered, v)
+		}
+	}
+	return filtered
+}
